requirements.txt
click
google-cloud-monitoring